split CoroutineUtils into several files
