--- unity-renderer/Assets/Scripts/TemporaryForMetaverse.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 023040a357298924ea8f20942390817e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
